Revert "Migration to V7"
This reverts commit 98635af2

+++ modules/web/src/eu/crewbase/lineup/web/traveloption/TravelOptionBookTransfer.java
@@ -54,7 +54,7 @@ public class TravelOptionBookTransfer extends AbstractEditor<TravelOption> {
 	}
 
     public void bookSeats(Component source) {
-    	travelOptionService.bookSeats(getItem().getId(), (Integer) bookedSeats.getValue());
+    	travelOptionService.bookSeats(getItem().getId(), bookedSeats.getValue());
     	this.close(WINDOW_CLOSE);
     }
